// interfaces/interfaces.d.ts
export interface JWTResponse {
    accessToken: string;
    refreshToken?: string;
    tokenType: string;
}

export interface ChangePasswordRequest {
    oldPassword: string;
    newPassword: string;
}

export interface GoogleLoginRequest {
    idToken: string;
}

export interface UserDTO {
    id?: number;
    email?: string;
    username: string;
    password?: string;
    profilePhotoUrl?: string;
}